Implemented Collectables, Objective Handler with edge case handling, and Level Scripting
I wanted to create a game/tool that allows developers to drop the entities they need into a scene and script out gameplay completely outside of the rest of the game's classes.

My eventual wish is to be able to script a whole level, akin to the way I've been able to in the Call of Duty Radiant and CS:GO Source Engines. Working towards that point by building nice, modular, decoupled code. This is a learning experience, as well as a reaffirmation of what I know right now.