Trying new things out with Cinemachine. Hopefully can integrate into my workflow on this project
I felt as though this was the right step to take as it'd help obfuscate a lot of the issues I was having with camera. Also, I only decided to hand code my camera solution because I was reenforcing my knowledge of 3D Math. Now that I'm sure of my strengths, I can have this handled with tools.